move some dependencies to dev, remove unused code and update export scripts and readme
