+++ bulid-export-mac.sh
@@ -23,6 +23,7 @@ echo -e "Kopierer filer til leveransemappen..."
 
 cp -r index.html $LEVERANSE_DIR
 cp -r package.json $LEVERANSE_DIR
+cp -r vite.config.js $LEVERANSE_DIR
 cp -r dist/style.css $LEVERANSE_DIR
 cp -r src $LEVERANSE_DIR/src
