+++ index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Workshop</title>
 
-    <link rel="stylesheet" property="stylesheet" media="screen,print" type="text/css"  href="./src/features/oversikt/style.css" />
    <link rel="stylesheet" property="stylesheet" media="screen,print" type="text/css"  href="./style.css" />
   </head>
   <body>
