+++ package.json
@@ -34,9 +34,7 @@
     "@sb1/ffe-tabs": "^13.0.3",
     "@sb1/ffe-tabs-react": "^8.0.3",
     "@sb1/ffe-webfonts": "^3.0.3",
-    "fs": "^0.0.1-security",
     "less": "^4.1.3",
-    "path": "^0.12.7",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -51,6 +49,7 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
+    "fs": "^0.0.1-security",
     "path": "^0.12.7",
     "prettier": "^2.8.8",
     "vite": "^4.3.9"
