--- src/Kort.jsx
@@ -1,63 +1,0 @@
-export const Kort = () => (
-  <svg width="62" height="62" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#a)">
-      <path
-        d="m46.116 32.499-6.018-8.737-10.316 2.912c-.596 1.815-1.676 3.966-2.513 6.004-.287.672.044 1.479.705 1.748.242.112.485.179.64.201 1.873.157 3.218-2.576 3.923-4.682.706.247 1.477.448 2.227.493 1.366.112 2.953.045 3.99-.963.595 2.755.617 3.45 3.108 4.839"
-        fill="#7EB5D2"
-      />
-      <path
-        d="M40.12 28.758c-.176-.247-.352-.516-.463-.83-.022-.089-.066-.156-.088-.223-.022-.067-.044-.135-.088-.202-.022-.067-.066-.134-.11-.224l-.089-.201c-.044-.09-.066-.202-.088-.314a2.972 2.972 0 0 1-.132-.605c0-.045-.066-.067-.132-.045-.066 0-.11.045-.11.09.022.224.044.425.088.65l.066.335.044.224c.022.09.044.157.066.247.022.067.066.157.088.224.044.067.066.157.133.224.176.291.33.582.484.851.023.045.067.09.089.135.198.246.374.515.573.74.22.245.44.514.705.716.088.067.199.157.309.224.154.112.33.224.485.336.11.09.242.157.352.224l.464.269c.044.022.088-.023.11-.067.022-.045.022-.112 0-.135-.155-.09-.309-.179-.441-.268a4.285 4.285 0 0 1-.772-.605c-.088-.09-.176-.157-.264-.247-.22-.224-.463-.448-.684-.672a4.207 4.207 0 0 1-.55-.695c.021-.067 0-.112-.045-.156Z"
-        fill="#002776"
-      />
-      <path
-        d="M48.916 35.456c-3.748-1.814-6.084-5.377-7.495-9.275-1.698.65-4.916 1.053-7.076.65-2.866 2.621-5.335 3.07-7.076 1.747-.86-.65-.772-1.993.176-2.531 1.741-.986 3.152-2.084 4.541-3.652.309-.336.727-.56 1.19-.605 4.123-.38 7.319-1.075 9.347-1.478a2.71 2.71 0 0 1 2.58.918c3.35 4.033 6.216 8.737 9.434 12.232.176.18 0 .47-.242.426l-5.38 1.568Z"
-        fill="#7EB5D2"
-      />
-      <path
-        d="M39.106 26.584c-1.102.202-2.27.292-3.395.224-.154-.022-.308-.022-.44-.044-.155-.023-.287-.023-.42-.045l-.462-.067h-.044c-.022 0-.022.022-.044.022l-.067.067-.088.09-.154.134c-.22.202-.44.404-.661.583-.22.18-.42.336-.64.493-.33.224-.66.425-1.014.582-.661.336-1.366.56-2.094.605-.132 0-.22.067-.22.112 0 .067.11.112.242.112.772-.045 1.521-.269 2.205-.582a6.802 6.802 0 0 0 1.036-.583c.242-.157.463-.313.683-.47.242-.18.485-.381.705-.56.089-.09.243-.224.22-.202a1.9 1.9 0 0 0 .354.045c.154.022.286.022.44.045.155 0 .31.022.463.022a19.686 19.686 0 0 0 3.461-.246c.199-.023.397-.068.596-.113.264-.067.55-.111.815-.201.132-.045.265-.067.397-.112l.198-.067.155-.045c.022 0 .022-.023.022 0l.044.09.154.403c.11.269.199.515.309.784.463 1.098 1.036 2.195 1.72 3.181.131.18.264.359.396.538.132.179.264.358.419.537.44.56.948 1.053 1.477 1.546.374.336.771.672 1.19.964.265.179.53.38.816.537.132.09.286.18.419.247l.44.224c.11.067.22.067.265 0 .044-.067-.044-.157-.154-.202l-.42-.224c-.131-.067-.264-.157-.418-.246-.287-.157-.529-.359-.794-.538a9.385 9.385 0 0 1-1.146-.963 13.255 13.255 0 0 1-1.389-1.569l-.396-.537c-.133-.18-.243-.359-.375-.538a26.748 26.748 0 0 1-1.698-3.114c-.11-.246-.22-.493-.33-.762l-.154-.38-.11-.292-.045-.09s0-.044-.022-.044l-.044.022-.176.067-.243.09-.198.067c-.133.045-.265.067-.397.112-.264.067-.53.135-.794.18-.198.044-.396.09-.595.111Z"
-        fill="#002776"
-      />
-      <path
-        d="M35.755 37.092c1.433-1.569 2.932-6.43 10.361-6.43 1.213 0 2.425.157 3.571.426-.022 0-2.711 11.627-13.931 6.004Z"
-        fill="#7EB5D2"
-      />
-      <path
-        d="m49.643 30.953-.397-.112-.198-.067c-.044-.022-.088-.022-.132-.045a.404.404 0 0 0-.154-.045 1.282 1.282 0 0 0-.31-.044c-.22 0-.418-.045-.616-.068a1.872 1.872 0 0 0-.618-.044c-.22 0-.44-.023-.66-.045-.177-.023-.376-.023-.574-.023-.154 0-.309 0-.441.023-.264.022-.529.044-.794.044h-.066c-.176 0-.374 0-.573.023-.132.022-.242.045-.375.067l-.242.067c-.176.045-.375.09-.551.112-.22.045-.441.09-.661.18a4.3 4.3 0 0 1-.574.179c-.198.067-.419.134-.595.246l-.22.112-.33.202c-.09.045-.155.09-.243.134-.11.045-.199.112-.287.18-.022.022 0 .067.022.112.022.044.066.09.088.067.11-.045.199-.112.309-.157.088-.045.154-.09.242-.112.11-.045.22-.09.353-.134l.22-.068a4.43 4.43 0 0 0 .574-.246c.176-.067.352-.157.55-.202.221-.045.442-.112.64-.157.176-.044.353-.112.53-.112l.242-.022c.132 0 .242-.022.352-.022h.088c.155-.023.31-.045.463-.09h.044c.265-.045.508-.045.772-.045h.992c.22 0 .44 0 .661.023.199.044.397.044.596.067.198 0 .418.022.595.09.11.022.198.044.308.067h.309l.198.022.397.067c.022 0 .044-.045.066-.112.044-.045.022-.09 0-.112Z"
-        fill="#002776"
-      />
-      <path
-        d="M47.615 32.745c-2.888-3.024-3.284-6.272-3.593-10.17-1.83 0-4.409-1.031-6.282-2.174-3.594 1.412-6.063.964-7.231-.896-.573-.918-.044-2.15 1.014-2.307 1.984-.314 3.681-.83 5.533-1.793a1.809 1.809 0 0 1 1.323-.156c3.196.873 5.599 1.41 7.473 1.77a5.152 5.152 0 0 1 3.88 3.337c1.697 4.683 3.46 9.208 5.092 13.04l-7.209-.65Z"
-        fill="#F8E9DD"
-      />
-      <path
-        d="m39.68 26.54-.662 1.814c-.264.695-1.036 1.053-1.72.785L11.86 19.438a1.365 1.365 0 0 1-.772-1.747L16.93 1.873c.264-.695 1.036-1.053 1.72-.784l25.416 9.7a1.366 1.366 0 0 1 .772 1.748l-3.24 8.737c-.794.426-1.61 1.098-2.073 2.308a3.932 3.932 0 0 0 .154 2.957Z"
-        fill="#002776"
-      />
-      <path
-        d="m21.625 13.703-3.858-1.479a.81.81 0 0 1-.462-1.03l.97-2.622a.789.789 0 0 1 1.014-.47l3.857 1.478a.809.809 0 0 1 .463 1.031l-.97 2.644c-.176.403-.617.604-1.014.448Z"
-        fill="#DC8000"
-      />
-      <path
-        d="M18.936 11.776c-.044.135-.088.247-.11.38l-.177.539c-.11.336-.22.672-.352 1.008-.022.067 0 .134.066.157.066.022.11 0 .132-.068.132-.336.265-.672.397-.985l.198-.516c.044-.112.11-.246.155-.358l.044-.09c.154-.515.352-1.03.55-1.546l.111-.268.044-.18c.11-.313.198-.627.309-.94a8.19 8.19 0 0 1 .242-.673c.11-.269.199-.56.309-.851.022-.067 0-.112-.066-.134-.067-.023-.11 0-.133.067-.11.291-.22.56-.33.851l-.265.672c-.11.314-.242.628-.375.919l-.066.157-.088.268c-.176.516-.353 1.031-.573 1.546l-.022.045ZM20.721 12.448c-.044.135-.088.247-.11.381l-.176.538c-.11.336-.22.672-.353 1.008-.022.067 0 .134.066.157.067.022.11 0 .133-.068a45.5 45.5 0 0 1 .396-.985l.199-.516c.044-.112.11-.246.154-.358l.044-.09c.154-.515.353-1.03.551-1.545l.11-.27.045-.178c.11-.314.198-.628.308-.941.066-.224.155-.449.243-.673.11-.268.198-.56.308-.85.022-.068 0-.113-.066-.135-.066-.023-.11 0-.132.067-.11.291-.22.56-.33.851l-.265.672c-.11.314-.243.628-.375.919l-.066.157-.088.268a21.92 21.92 0 0 1-.573 1.546l-.023.045Z"
-        fill="#002776"
-      />
-      <path
-        d="M22.375 10.633a.309.309 0 0 1-.133-.044l-.198-.068c-.132-.044-.242-.067-.375-.112-.022 0-.066.045-.088.09-.022.067-.022.112 0 .112.11.045.243.112.353.157l.176.09.133.067.022.022c.198.045.374.112.55.202l.089.045.066.022c.11.022.242.067.353.09.088.022.154.044.242.09.11.044.198.066.309.111.022 0 .066-.045.066-.09.022-.067.022-.111 0-.111-.088-.045-.199-.09-.309-.135-.088-.022-.154-.067-.242-.09-.11-.044-.22-.111-.33-.156l-.067-.023-.044-.044c-.176-.068-.375-.113-.551-.225h-.022ZM18.23 9.065a.31.31 0 0 1-.132-.045l-.198-.067c-.133-.045-.243-.067-.375-.112-.022 0-.066.045-.088.09-.022.067-.022.112 0 .112.11.045.242.112.353.157l.176.09.132.066.022.023c.199.045.375.112.551.201l.088.045.067.023c.11.022.242.067.352.09.088.022.155.044.243.089.11.045.198.067.309.112.021 0 .066-.045.066-.09.022-.067.022-.112 0-.112a4.287 4.287 0 0 0-.309-.134c-.088-.023-.154-.067-.242-.09-.11-.045-.22-.112-.331-.157l-.066-.022-.022-.045c-.177-.067-.375-.112-.551-.224h-.044ZM21.89 11.978c-.044 0-.088-.023-.132-.045l-.199-.067c-.132-.045-.242-.067-.375-.112-.022 0-.066.045-.088.09-.022.067-.022.111 0 .111.11.045.243.112.353.157l.176.09.133.067.022.022c.198.045.374.112.55.202l.089.045.066.022c.11.023.242.067.353.09.088.022.154.045.242.09.11.044.199.067.309.111.022 0 .066-.044.066-.09 0-.044.022-.111 0-.111a4.276 4.276 0 0 0-.309-.135c-.088-.022-.154-.067-.242-.09-.11-.044-.22-.111-.33-.156l-.067-.023-.044-.044c-.176-.068-.375-.112-.551-.224h-.022ZM17.745 10.387c-.044 0-.088-.022-.132-.045l-.198-.067c-.132-.045-.243-.067-.375-.112-.022 0-.066.045-.088.09-.022.067-.022.112 0 .112.11.044.242.112.353.156l.176.09.132.067.022.023c.199.044.375.112.551.201l.088.045.067.023c.11.022.242.067.352.09.089.021.155.044.243.089.11.045.198.067.308.112.023 0 .067-.045.067-.09s.022-.112 0-.112a4.292 4.292 0 0 0-.309-.134c-.088-.023-.154-.068-.242-.09-.11-.045-.22-.112-.331-.157l-.066-.022-.088-.023c-.177-.067-.375-.112-.552-.224l.023-.022Z"
-        fill="#002776"
-      />
-      <path
-        d="M49.687 99.933c-.044-.135-4.475-24.554-6.966-40.281l11.992-24.24a2.546 2.546 0 0 0-.793-3.182l-5.401-3.764c-.992.65-3.527 2.42-5.246 6.072-.772 1.658-2.447 2.733-4.255 2.576-.066 0-.154-.022-.22-.022a1.294 1.294 0 0 1-.97-.65c-.794-1.3-1.477-3.473-1.477-6.183v-.068a15.04 15.04 0 0 0-.86-5.04 77.809 77.809 0 0 0-1.08-2.846c-.33-.806-1.125-1.299-1.984-1.21-.22.023-.463.068-.683.135-.75.202-1.301.829-1.433 1.613-.11.74-.22 1.501-.287 2.263-.22 2.061-.022 4.144.617 6.093.441 1.322.772 2.913.772 4.75 0 5.489-3.395 11.605-3.395 17.452 0 2.89.794 5.198 2.623 7.864.22 8.02.595 16.04.595 24.06 0 4.145-.11 8.312-.374 12.457"
-        fill="#F8E9DD"
-      />
-      <path
-        d="M29.98 23.538a13.236 13.236 0 0 0-.287 3.091c0 .27.022.538.044.807.023.134.023.269.045.403l.044.403c.044.381.11.74.176 1.12.066.404.154.807.242 1.21.155.717.375 1.457.574 2.308a13.7 13.7 0 0 1 .418 2.307c.023.381.045.785 0 1.165-.022.202-.022.381-.044.583l-.066.582c-.176 1.501-.55 3.025-.904 4.503-.418 1.636-.925 3.271-1.366 4.884-.353 1.367-.706 2.8-.86 4.234-.044.56-.132 1.143-.132 1.703-.022.583-.022 1.143.022 1.725.044.493.11.986.198 1.479.11.493.243.986.397 1.456.176.47.353.941.573 1.412.22.448.441.896.706 1.321.066.135.154.27.242.381.088.157.176.291.287.448l.154.224.022.023V61.422c0 .358.022.717.022 1.075l.066 2.128.11 2.846c.022.604.044 1.232.088 1.859.067 1.411.089 2.845.133 4.234.044 1.68.066 3.428.132 5.13.044 1.502.066 3.025.066 4.504v6.676c0 .963-.022 1.904-.044 2.867a64.547 64.547 0 0 1-.088 2.061c-.044.852-.088 1.748-.132 2.622 0 .179.022.313.088.336.066 0 .11-.135.132-.314.066-.874.11-1.77.154-2.621.022-.695.067-1.367.089-2.061l.11-2.89c.022-.605.044-1.21.044-1.815.044-1.635.022-3.27 0-4.884l-.066-4.503c0-1.725-.066-3.472-.133-5.153-.044-1.389-.11-2.822-.154-4.256l-.044-1.86-.066-2.823c0-.224-.022-.448-.022-.672l-.066-1.814-.022-.448V61.175l-.022-.045-.265-.38c-.088-.112-.154-.247-.242-.359a1.943 1.943 0 0 1-.22-.358 14.732 14.732 0 0 1-.706-1.3c-.22-.448-.397-.896-.551-1.366a10.51 10.51 0 0 1-.375-1.434 15.38 15.38 0 0 1-.198-1.456c-.045-.56-.045-1.12-.023-1.68 0-.56.089-1.12.133-1.68.176-1.412.485-2.846.837-4.19.397-1.613.882-3.249 1.323-4.862.441-1.5.772-3.047.948-4.57l.066-.582c0-.202.022-.404.022-.605a7.397 7.397 0 0 0-.022-1.21c-.066-.807-.22-1.59-.375-2.375-.176-.851-.374-1.59-.55-2.307-.11-.404-.2-.785-.287-1.188-.066-.358-.155-.717-.199-1.098l-.066-.403c-.022-.134-.022-.269-.044-.403-.022-.269-.066-.538-.066-.807a12.635 12.635 0 0 1 .22-3.024c.022-.18 0-.336-.044-.336-.066.067-.132.201-.176.38ZM36.395 27.1c-.044-.18-.066-.359-.11-.538a9.52 9.52 0 0 0-.177-.762 8.067 8.067 0 0 0-.507-1.389c-.044-.09-.11-.112-.154-.09-.044.023-.066.113-.044.18.198.448.33.919.44 1.389.067.247.11.493.155.74.044.179.044.358.066.515 0 .044 0 .09.022.134.132.74.154 1.501.154 2.218v.627c.023.448.023.919.044 1.39.023.335.045.671.089.985.044.403.132.829.242 1.232.022.09.088.135.154.112.067-.022.089-.09.067-.179-.089-.403-.155-.829-.22-1.232-.045-.314-.067-.65-.089-.964-.022-.448 0-.896.022-1.366v-.627a40.871 40.871 0 0 0-.132-2.286s0-.044-.022-.09ZM43.14 61.108c-.088-.47-.154-.94-.242-1.434.573-1.142 1.124-2.307 1.697-3.45a260.621 260.621 0 0 1 1.764-3.517c.771-1.524 1.433-2.89 2.072-4.19.353-.717.705-1.456 1.058-2.173.33-.672.64-1.322.97-1.994l1.41-2.867c.883-1.815 1.764-3.63 2.668-5.444.154-.314.243-.605.176-.628-.044-.022-.22.202-.374.516-.904 1.814-1.808 3.63-2.734 5.444l-1.41 2.822-.993 1.972c-.352.717-.727 1.434-1.08 2.15-.639 1.3-1.322 2.667-2.072 4.213-.573 1.165-1.146 2.352-1.72 3.517-.594 1.187-1.168 2.375-1.763 3.54.088.515.154 1.03.243 1.545l.242 1.569c.485 2.934.97 5.914 1.455 8.804.507 3.114 1.058 6.34 1.631 9.5.22 1.186.42 2.351.64 3.539.33 1.86.683 3.719 1.014 5.6l.727 4.033c.309 1.68.617 3.406.948 5.108.066.336.176.627.22.605.044-.022.044-.291 0-.65-.308-1.702-.617-3.45-.925-5.13a257.79 257.79 0 0 1-.706-4.033l-.97-5.623-.595-3.517c-.53-3.181-1.102-6.385-1.631-9.5-.507-2.89-1.014-5.869-1.477-8.781l-.243-1.546ZM50.437 22.82c.44 1.143.904 2.33 1.323 3.518.242.65.462 1.21.66 1.747l.376.897.352.829.507 1.187.97 2.218c.066.134.154.224.199.201.044-.022.044-.156 0-.29-.309-.74-.618-1.502-.926-2.264l-.463-1.187-.309-.829c-.11-.291-.22-.605-.33-.896-.199-.538-.441-1.098-.684-1.747a51.515 51.515 0 0 1-1.323-3.495l-.22-.605c-.22-.56-.419-1.12-.617-1.68a5.386 5.386 0 0 0-.86-1.613 5.532 5.532 0 0 0-1.455-1.322 4.626 4.626 0 0 0-1.852-.672l-1.388-.247c-.728-.134-1.455-.291-2.183-.448-.529-.112-1.036-.246-1.565-.38a86.71 86.71 0 0 1-1.984-.516c-.132-.045-.264-.022-.264.045-.022.067.088.134.22.18.661.2 1.323.38 1.984.537.53.134 1.036.269 1.565.403.728.18 1.455.359 2.183.515.463.112.925.202 1.366.292a5.125 5.125 0 0 1 3.087 1.859c.352.448.639.941.815 1.501.199.56.397 1.12.617 1.68l.199.583Z"
-        fill="#002776"
-      />
-    </g>
-    <defs>
-      <clipPath id="a">
-        <path fill="#fff" d="M0 0h62v62H0z" />
-      </clipPath>
-    </defs>
-  </svg>
-);
