--- src/components/ExampleOverview/AccountDetails.jsx
@@ -1,7 +1,0 @@
-import { useParams } from "react-router-dom";
-
-export const AccountDetails = () => {
-  let { accountNumber } = useParams();
-
-  return <>Kontoinfo for {accountNumber}</>;
-};
