--- src/components/ExampleOverview/kontoliste.jsx
@@ -1,58 +1,0 @@
-export const kontoliste = {
-  accounts: [
-    {
-      accountNumber: "42133900246",
-      name: "BRUKSKONTO 1",
-      description: "BRUKSKONTO FRI",
-      balance: 542.0,
-      availableBalance: 542.0,
-      currencyCode: "NOK",
-      type: "BRUK",
-    },
-    {
-      accountNumber: "42133900290",
-      name: "BRUKSKONTO 2",
-      description: "BRUKSKONTO FRI",
-      balance: 645.0,
-      availableBalance: 600.0,
-      currencyCode: "NOK",
-      type: "BRUK",
-    },
-    {
-      accountNumber: "42133901456",
-      name: "BRUKSKONTO 3",
-      description: "BRUKSKONTO FRI",
-      balance: 645.0,
-      availableBalance: 600.0,
-      currencyCode: "NOK",
-      type: "BRUK",
-    },
-    {
-      accountNumber: "42133909846",
-      name: "Sparekonto",
-      description: "Min sparekonto",
-      balance: 10000.0,
-      availableBalance: 10000.0,
-      currencyCode: "NOK",
-      type: "SAVINGS",
-    },
-    {
-      accountNumber: "42132909890",
-      name: "BSU",
-      description: "Min BSU-konto",
-      balance: 10000.0,
-      availableBalance: 10000.0,
-      currencyCode: "NOK",
-      type: "BSU",
-    },
-    {
-      accountNumber: "42133906559",
-      name: "Lån",
-      description: "Mitt lån",
-      balance: -4000.0,
-      availableBalance: -4000.0,
-      currencyCode: "NOK",
-      type: "LOAN",
-    },
-  ],
-};
