+++ src/main.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./App.less";
+import "./App.less"; // TODO: Fjern når koden lever i codesandbox
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import App from "./App";
-import Navigation from "./components/navigation.jsx";
+import Navigation from "./navigation.jsx";
 import SlikSkalDetSeUt from "./features/slik-skal-det-se-ut/SlikSkalDetSeUt.jsx";
 import Banken from "./features/banken/Banken.jsx";
